fix data importing tests

+++ cypress/e2e/data-import/importing-marc-authority-files.cy.js
@@ -11,7 +11,7 @@ import Logs from '../../support/fragments/data_import/logs/logs';
 import SettingsMenu from '../../support/fragments/settingsMenu';
 import NewJobProfile from '../../support/fragments/data_import/job_profiles/newJobProfile';
 
-describe('Importing MARC Authority files', () => {
+describe('Data Import -> Importing MARC Authority files', () => {
   const testData = {};
   const jobProfileToRun = 'Default - Create SRS MARC Authority';
   const createdJobProfile = {
@@ -46,7 +46,7 @@ describe('Importing MARC Authority files', () => {
   });
 
   afterEach('Deleting data', () => {
-    MarcAuthority.deleteViaAPI(createdAuthorityID);
+    if(createdAuthorityID) MarcAuthority.deleteViaAPI(createdAuthorityID);
 
     cy.loginAsAdmin({ path: TopMenu.dataImportPath, waiter: DataImport.waitLoading });
     DataImport.selectLog();
